uint64_t is not size_t on OSX (OSX)

+++ builtin/merged_dataset.cc
@@ -594,7 +594,7 @@ struct MergedDataset::Itl
         return result;
     }
 
-    virtual size_t getRowCount() const
+    virtual uint64_t getRowCount() const
     {
         return rowIndex.size();
     }
