+++ builtin/transposed_dataset.cc
@@ -298,7 +298,7 @@ struct TransposedDataset::Itl
         return result;
     }
 
-    virtual size_t getRowCount() const
+    virtual uint64_t getRowCount() const
     {
         return dataset->getFlattenedColumnCount();
     }
